backend/requirements.txt
Django==6.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.5.1
psycopg==3.1.12
supabase==2.4.1
drf-nested-routers==0.93.4
python-dotenv==1.0.0
django-cors-headers==4.3.1
gunicorn==21.2.0
dj-database-url==2.1.0
Pillow==11.0.0
requests==2.31.0
python-multipart==0.0.6